python/requirements.txt
langchain
python-dotenv
openai
sqlalchemy
pymssql